% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{compare_groups}
\alias{compare_groups}
\title{Compare group expression data (statistical test)}
\usage{
compare_groups(
  data,
  group1_cols,
  group2_cols,
  group1_name = "Group1",
  group2_name = "Group2",
  test_type = "t-test",
  adjust_method = "fdr"
)
}
\arguments{
\item{data}{A dataframe where rows represent genes, and columns represent samples.}

\item{group1_cols}{A character vector specifying the column names for samples belonging to the first group.}

\item{group2_cols}{A character vector specifying the column names for samples belonging to the second group.}

\item{group1_name}{A string representing the name of the first group. This will be used in the result as the mean column name.
Default is \code{"Group1"}.}

\item{group2_name}{A string representing the name of the second group. This will be used in the result as the mean column name.
Default is \code{"Group2"}.}

\item{test_type}{A string indicating the statistical test to use. Options are \code{"t-test"} for a t-test or \code{"wilcox"} for
a Wilcoxon rank-sum test. Defaults to \code{"t-test"}.}

\item{adjust_method}{A string specifying the p-value adjustment method to account for multiple testing. Defaults to \code{"fdr"}.
See \code{p.adjust} function for available methods.}
}
\value{
A data.frame with p.value, p.adjusted, and the mean value of each group.
}
\description{
This function compares the expression values of two groups using statistical tests (t-test or Wilcoxon rank-sum test).
It calculates the mean expression values for each group, performs the specified statistical test,
and adjusts the resulting p-values for multiple testing.
}
\examples{
# example code
mean_expression_df <- data.frame(
  AITL_1 = c(1.2, 3.4, 5.6),
  AITL_2 = c(1.5, 3.7, 5.9),
  TFH_1 = c(2.1, 3.5, 4.6),
  TFH_2 = c(2.4, 3.8, 4.9),
  row.names = c("Gene1", "Gene2", "Gene3")
)

# Define group columns
AITL_samples <- grep("^AITL", colnames(mean_expression_df), value = TRUE)
TFH_samples <- grep("^TFH", colnames(mean_expression_df), value = TRUE)

# Perform comparison
compare_groups(mean_expression_df, AITL_samples, TFH_samples,group1_name='AITL',group2_name='Healthy', test_type = "t-test", adjust_method = "fdr")


}
