% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Heatmap with Pheatmap}
\usage{
create_heatmap(
  data,
  filename = "heatmap_plot.png",
  save = TRUE,
  scale = "row",
  fontsize = 7,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  show_rownames = FALSE,
  show_colnames = TRUE,
  annotation = NULL,
  title = "Heatmap",
  top = NULL
)
}
\arguments{
\item{data}{Data frame with normalized values and genes in rownames}

\item{filename}{Name of the file to save the plot}

\item{save}{Whether to save the plot or not (Default TRUE).}

\item{scale}{Scale data (Default rows).}

\item{fontsize}{Size of the text}

\item{cluster_rows}{Default True}

\item{cluster_cols}{Default True}

\item{show_rownames}{Default False}

\item{show_colnames}{Default True}

\item{annotation}{Data frame with annotaion}

\item{title}{Title of the plot.}

\item{top}{Select top rows (and do the heatmap with them).}
}
\description{
Heatmap with Pheatmap
}
\examples{
data(data_heatmap)
data(pD_annotation)
create_heatmap(data_heatmap,top=10,save = FALSE,annotation = pD_annotation,show_rownames = TRUE)
}
