% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_volcano.R
\name{enhancedvolcanoplot}
\alias{enhancedvolcanoplot}
\title{Volcano Plot with EnhancedVolcano}
\usage{
enhancedvolcanoplot(
  top.table,
  filename = "volcano_plot.png",
  save = TRUE,
  title = "Group1 vs Group2",
  top_genes = "NULL"
)
}
\arguments{
\item{top.table}{Results from the DE analysis (logFC,P.value,), with genes in rownames.}

\item{filename}{Filename to save the plot (Default volcano_plot.png).}

\item{save}{Whether to save the plot or not (Default TRUE).}

\item{title}{Title of the plot.}

\item{top_genes}{Vector containing names of genes to highlight in the plot.}
}
\description{
This function creates an enhanced volcano plot for visualizing
differential expression results, such as log-fold changes and p-values.
}
\examples{
# Example data
data(top.table)
enhancedvolcanoplot(top.table, title = "Example Volcano Plot",save = FALSE)

}
